Modernize Jazz++ MIDI Sequencer for Linux
- Added root CMakeLists.txt with modern C++17 support and wxWidgets 3.2+
- Upgraded string handling for wxWidgets 3.2+ Unicode standards across 15+ files
- Fixed memory bugs (mismatched new[] / delete in Track.cpp and Synth.cpp)
- Fixed missing break statements in SysEx player event handler (Player.cpp)
- Enhanced ALSA PCM audio device negotiation with set_channels_near / set_rate_near
- Cleaned up compiler warnings and added desktop integration (.desktop & .gitignore)
- Zero-warning build verified on GCC 13.3 / Ubuntu 24.04 / Zorin OS 18

+++ src/Random.cpp
@@ -105,9 +105,9 @@ double JZRndArray::operator[](double f)
   {
     i = 0;
   }
-  else if (i >= mArray.size() - 2)
+  else if (i >= (int)mArray.size() - 2)
   {
-    i = mArray.size() - 2;
+    i = (int)mArray.size() - 2;
   }
   JZMapper Map(i, i + 1, mArray[i], mArray[i + 1]);
   return Map.XToY(f);
@@ -161,7 +161,7 @@ int JZRndArray::Random(double RandonValue)
   }
   i--;
 
-  assert(i >= 0 && i < mArray.size());
+  assert(i >= 0 && (size_t)i < mArray.size());
   return i;
 }
